H1 tri element fem solves weakform.
